+++ pkg/manifests/manifests.go
@@ -33,6 +33,7 @@ import (
 	"github.com/openshift/cluster-monitoring-operator/pkg/promqlgen"
 	"github.com/pkg/errors"
 	monv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+	yaml2 "gopkg.in/yaml.v2"
 	admissionv1 "k8s.io/api/admissionregistration/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
@@ -242,6 +243,9 @@ var (
 	AuthProxyRedirectURLFlag  = "-redirect-url="
 
 	TrustedCABundleKey = "ca-bundle.crt"
+
+	AdditionalAlertmanagerConfigSecretKey  = "alertmanager-configs.yaml"
+	AdditionalAlertmanagerConfigSecretName = "prometheus-k8s-additional-alertmanager-configs"
 )
 
 type Factory struct {
@@ -1352,7 +1356,148 @@ func (f *Factory) PrometheusK8s(host string, grpcTLS *v1.Secret, trustedCABundle
 		}
 	}
 
+	if f.config.ClusterMonitoringConfiguration.PrometheusK8sConfig.AlertmanagerConfigs != nil {
+		p.Spec.AdditionalAlertManagerConfigs = &v1.SecretKeySelector{
+			Key: AdditionalAlertmanagerConfigSecretKey,
+			LocalObjectReference: v1.LocalObjectReference{
+				Name: AdditionalAlertmanagerConfigSecretName,
+			},
+		}
+		p.Spec.Secrets = append(p.Spec.Secrets, getAdditionalAlertmanagerSecrets(f.config.ClusterMonitoringConfiguration.PrometheusK8sConfig.AlertmanagerConfigs)...)
+	}
+
 	return p, nil
+}
+
+func (f *Factory) AdditionalAlertManagerConfigsSecret() (*v1.Secret, error) {
+	alertmanagerConfigs := f.config.ClusterMonitoringConfiguration.PrometheusK8sConfig.AlertmanagerConfigs
+	if len(alertmanagerConfigs) == 0 {
+		return nil, nil
+	}
+
+	cfgs := []yaml2.MapSlice{}
+	for _, alertmanagerConfig := range alertmanagerConfigs {
+		cfg := yaml2.MapSlice{}
+
+		if alertmanagerConfig.Scheme != "" {
+			cfg = append(cfg, yaml2.MapItem{
+				Key: "scheme", Value: alertmanagerConfig.Scheme,
+			})
+		}
+		if alertmanagerConfig.PathPrefix != "" {
+			cfg = append(cfg, yaml2.MapItem{
+				Key: "path_prefix", Value: alertmanagerConfig.PathPrefix,
+			})
+		}
+		if alertmanagerConfig.APIVersion != "" {
+			cfg = append(cfg, yaml2.MapItem{
+				Key: "api_version", Value: alertmanagerConfig.APIVersion,
+			})
+		}
+		if alertmanagerConfig.Timeout != nil {
+			cfg = append(cfg, yaml2.MapItem{
+				Key: "timeout", Value: alertmanagerConfig.Timeout,
+			})
+		}
+
+		auth := yaml2.MapSlice{}
+		if alertmanagerConfig.BearerToken != nil {
+			if alertmanagerConfig.BearerToken.Name == "" {
+				return nil, errors.Errorf("secret %q for bearer token not found", alertmanagerConfig.BearerToken.Name)
+			}
+			if alertmanagerConfig.BearerToken.Key == "" {
+				return nil, errors.Errorf("secret key %q for bearer token not found", alertmanagerConfig.BearerToken.Key)
+			}
+			auth = append(auth, yaml2.MapItem{
+				Key: "credentials_file", Value: fmt.Sprintf("/etc/prometheus/secrets/%s/%s",
+					alertmanagerConfig.BearerToken.Name, alertmanagerConfig.BearerToken.Key),
+			})
+		}
+		if len(auth) != 0 {
+			cfg = append(cfg, yaml2.MapItem{
+				Key:   "authorization",
+				Value: auth,
+			})
+		}
+		tlsConfig := yaml2.MapSlice{}
+		if alertmanagerConfig.TLSConfig.CA != nil {
+			if alertmanagerConfig.TLSConfig.CA.Name == "" {
+				return nil, errors.Errorf("secret %q for ca not found", alertmanagerConfig.TLSConfig.CA.Name)
+			}
+			if alertmanagerConfig.TLSConfig.CA.Key == "" {
+				return nil, errors.Errorf("secret key %q for ca not found", alertmanagerConfig.TLSConfig.CA.Key)
+			}
+			tlsConfig = append(tlsConfig, yaml2.MapItem{
+				Key: "ca_file", Value: fmt.Sprintf("/etc/prometheus/secrets/%s/%s",
+					alertmanagerConfig.TLSConfig.CA.Name, alertmanagerConfig.TLSConfig.CA.Key),
+			})
+		}
+		if alertmanagerConfig.TLSConfig.Cert != nil {
+			if alertmanagerConfig.TLSConfig.Cert.Name == "" {
+				return nil, errors.Errorf("secret %q for cert not found", alertmanagerConfig.TLSConfig.Cert.Name)
+			}
+			if alertmanagerConfig.TLSConfig.Cert.Key == "" {
+				return nil, errors.Errorf("secret key %q for cert not found", alertmanagerConfig.TLSConfig.Cert.Key)
+			}
+			tlsConfig = append(tlsConfig, yaml2.MapItem{
+				Key: "cert_file", Value: fmt.Sprintf("/etc/prometheus/secrets/%s/%s",
+					alertmanagerConfig.TLSConfig.Cert.Name, alertmanagerConfig.TLSConfig.Cert.Key),
+			})
+		}
+		if alertmanagerConfig.TLSConfig.Key != nil {
+			if alertmanagerConfig.TLSConfig.Key.Name == "" {
+				return nil, errors.Errorf("secret %q for cert key not found", alertmanagerConfig.TLSConfig.Key.Name)
+			}
+			if alertmanagerConfig.TLSConfig.Key.Key == "" {
+				return nil, errors.Errorf("secret key %q for cert key not found", alertmanagerConfig.TLSConfig.Key.Key)
+			}
+			tlsConfig = append(tlsConfig, yaml2.MapItem{
+				Key: "key_file", Value: fmt.Sprintf("/etc/prometheus/secrets/%s/%s",
+					alertmanagerConfig.TLSConfig.Key.Name, alertmanagerConfig.TLSConfig.Key.Key),
+			})
+		}
+		if alertmanagerConfig.TLSConfig.ServerName != "" {
+			tlsConfig = append(tlsConfig, yaml2.MapItem{
+				Key: "server_name", Value: alertmanagerConfig.TLSConfig.ServerName,
+			})
+		}
+		tlsConfig = append(tlsConfig, yaml2.MapSlice{
+			{Key: "insecure_skip_verify", Value: alertmanagerConfig.TLSConfig.InsecureSkipVerify},
+		}...)
+
+		cfg = append(cfg, yaml2.MapItem{
+			Key:   "tls_config",
+			Value: tlsConfig,
+		})
+
+		if len(alertmanagerConfig.StaticConfigs) > 0 {
+			sc := yaml2.MapSlice{
+				{Key: "targets", Value: alertmanagerConfig.StaticConfigs},
+			}
+			cfg = append(cfg, yaml2.MapItem{
+				Key:   "static_configs",
+				Value: []yaml2.MapSlice{sc},
+			})
+			cfgs = append(cfgs, cfg)
+		}
+	}
+
+	amConfigYaml, err := yaml2.Marshal(cfgs)
+	if err != nil {
+		return nil, err
+	}
+
+	additionalPromToAmConfigSecret := &v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      AdditionalAlertmanagerConfigSecretName,
+			Namespace: f.namespace,
+		},
+		Data: map[string][]byte{
+			AdditionalAlertmanagerConfigSecretKey: amConfigYaml,
+		},
+	}
+
+	return additionalPromToAmConfigSecret, nil
 }
 
 func (f *Factory) PrometheusUserWorkload(grpcTLS *v1.Secret) (*monv1.Prometheus, error) {
@@ -3584,4 +3729,47 @@ func trustedCABundleVolume(configMapName, volumeName string) v1.Volume {
 			},
 		},
 	}
+}
+
+func getAdditionalAlertmanagerSecrets(alertmanagerConfigs []AdditionalAlertmanagerConfig) []string {
+
+	secretsName := []string{}
+	for _, alertmanagerConfig := range alertmanagerConfigs {
+		if alertmanagerConfig.TLSConfig.CA != nil {
+			secretsName = append(secretsName, alertmanagerConfig.TLSConfig.CA.Name)
+		}
+		if alertmanagerConfig.TLSConfig.Cert != nil {
+			secretsName = append(secretsName, alertmanagerConfig.TLSConfig.Cert.Name)
+		}
+		if alertmanagerConfig.TLSConfig.Key != nil {
+			secretsName = append(secretsName, alertmanagerConfig.TLSConfig.Key.Name)
+		}
+		if alertmanagerConfig.BearerToken != nil {
+			secretsName = append(secretsName, alertmanagerConfig.BearerToken.Name)
+		}
+	}
+	return removeEmptyDuplicates(secretsName)
+}
+
+func removeEmptyDuplicates(elements []string) []string {
+	// Use map to record duplicates as we find them.
+	encountered := map[string]struct{}{}
+	result := []string{}
+
+	for _, v := range elements {
+		if _, found := encountered[v]; found {
+			continue
+		}
+		encountered[v] = struct{}{}
+		if v != "" {
+			// Append to result slice if it is not empty
+			result = append(result, v)
+		}
+	}
+	// Return the new slice.
+	return result
+}
+
+func (f *Factory) GetAdditionalAlertmanagerConfigSecretName() string {
+	return AdditionalAlertmanagerConfigSecretName
 }